Rename patientFacilityTransfer package to facilityselection
