+++ app/src/main/java/org/simple/clinic/scheduleappointment/facilityselection/PatientFacilityChangeActivity.kt
@@ -1,4 +1,4 @@
-package org.simple.clinic.scheduleappointment.patientFacilityTransfer
+package org.simple.clinic.scheduleappointment.facilityselection
 
 import android.content.Context
 import android.content.Intent
